Refactor user function signature to "Rust" / "OCaml" syntax
From:
```
fun f a b : t -> t -> t =

end
```

to:

```
fn f (a : t, b : t) : t

end
```

The first syntax may lead to think that function might be applied partially, but in fact IMITATOR doesn't support partials applications. Indeed, call to a function in IMITATOR can be made like this : `f(a, b)`.

As we choose syntax close to OCaml, and we want a signature syntax that are more close conceptually of the way we call functions. I choose a syntax inspired by rust, because it's syntactically and conceptually the nearest language of OCaml that not supporting (directly) partials applications.