Implement first working version of AlgoPTG
This version follows the algorithm by Cassez et. al.
It has been tested heavily, but we still need to test it more -
especially for PTG with parameters.

Compiler bug: If run on sufficiently large PTG a segmentation fault or
Out_of_memory error might occur - this does not happen when compiling to
bytecode. We should switch to a newer version of ocaml
perhaps to prevent this bug...